requirements.txt
# ============================================================
# AI Decision Intelligence System — Python Dependencies v2
# Pin versions for reproducible enterprise builds.
# ============================================================

# ── Core data processing ────────────────────────────────────
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2

# ── Forecasting (Prophet is optional; linear fallback used if absent) ──
# prophet==1.1.5

# ── API framework ───────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1

# ── AWS SDK ─────────────────────────────────────────────────
boto3==1.34.110
botocore==1.34.110

# ── LLM integration (Anthropic Claude) ──────────────────────
anthropic>=0.28.0

# ── Reliability ─────────────────────────────────────────────
tenacity==8.3.0

# ── Config & environment ────────────────────────────────────
pyyaml==6.0.1
python-dotenv==1.0.1

# ── CLI reporting (rich terminal output) ────────────────────
rich>=13.7.0

# ── HTML reporting (interactive charts) ─────────────────────
plotly>=5.22.0

# ── Testing ─────────────────────────────────────────────────
pytest==8.2.0
pytest-cov==5.0.0
